requirements.txt
# Python 3.13: use 4.x connector and current snowflake-ml-python (1.7.x pins are not compatible with 3.13)
snowflake-connector-python==4.4.0
snowflake-ml-python==1.32.0
streamlit==1.40.2
pandas==2.2.3
numpy>=2.0,<3
faker==30.8.2
python-dotenv==1.0.1
plotly==5.24.1
